requirements.txt
pytest==6.2.4
requests==2.25.1
loguru==0.5.3

mysql-connector-python~=8.0.30